refactor: qc-application authentication 도메인 패키지 정리
