+++ qc-application/src/main/java/com/eager/questioncloud/application/api/authentication/dto/AuthenticationRequest.java
@@ -1,4 +1,4 @@
-package com.eager.questioncloud.application.api.authentication;
+package com.eager.questioncloud.application.api.authentication.dto;
 
 import com.eager.questioncloud.application.validator.EmailValidator;
 import com.eager.questioncloud.application.validator.PasswordValidator;
