+++ qc-application/src/main/java/com/eager/questioncloud/application/api/authentication/dto/AuthenticationToken.java
@@ -1,4 +1,4 @@
-package com.eager.questioncloud.application.api.authentication;
+package com.eager.questioncloud.application.api.authentication.dto;
 
 import lombok.Getter;
