+++ qc-application/src/main/java/com/eager/questioncloud/application/security/JwtAuthenticationFilter.java
@@ -1,6 +1,6 @@
 package com.eager.questioncloud.application.security;
 
-import com.eager.questioncloud.application.api.authentication.AuthenticationTokenManager;
+import com.eager.questioncloud.application.api.authentication.implement.AuthenticationTokenManager;
 import com.eager.questioncloud.core.domain.user.dto.UserWithCreator;
 import com.eager.questioncloud.core.domain.user.infrastructure.UserRepository;
 import jakarta.servlet.FilterChain;
