+++ qc-application/src/test/java/com/eager/questioncloud/application/api/authentication/AuthenticationProcessorTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
 
+import com.eager.questioncloud.application.api.authentication.dto.SocialAuthentication;
+import com.eager.questioncloud.application.api.authentication.implement.AuthenticationProcessor;
 import com.eager.questioncloud.core.domain.social.SocialAPIManager;
 import com.eager.questioncloud.core.domain.user.dto.CreateUser;
 import com.eager.questioncloud.core.domain.user.enums.AccountType;
